requirements.txt
streamlit
pymupdf
scikit-learn
numpy
google-generative-ai
rich
pickle
faiss-cpu
goole-api-core